Refactored a bit so that everything starts from the frontend
- Still needs more refactoring.

+++ client/src/main/java/frontend/ProgressBarPanel.java
@@ -10,10 +10,10 @@ public class ProgressBarPanel
 {
     private JTextArea textArea;
     private JProgressBar progressBar;
-    private int totalVideoTimeSeconds;
+    private double totalVideoTimeSeconds;
 
     
-    public ProgressBarPanel(int totalVideoTimeSeconds)
+    public ProgressBarPanel(double totalVideoTimeSeconds)
     {
     	this.totalVideoTimeSeconds = totalVideoTimeSeconds;
     	createProgressBarPanel();
